made cart react component
