+++ client/store/guest-cart.js
@@ -1,0 +1,5 @@
+// In this file we are going to write some functions that mimic the cart server routes
+// and store data into the browsers local storage Eg : delete product from cart, it takes the
+// product id, and deletes it from the local storage
+// We have to serialize it to JSON everytime we store it and deserialize it to read it
+//
